[ADDED] Primary Backend Setup
The backend for the project has been setup in Django

[chat]
The central project

[rooms]
The app that is responsible for the Room model, its views and URLs.

[users]
The app that is responsible for the User and Message model, views and URLs.

Additional
- routing.py in chat
The routing.py is used to route the WebSockets that are opened at the url /ws/rooms/[room-name].

- consumers.py in chat
The routing.py routes the WebSocket to the consumers.py. Essentially, the traffic that will move across the WebSockets is defined by the consumers.py